+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 1.x branch
 ## 1.1 branch
+### 1.1.13
+* added `is_html()` global function. This also provides the `assertIsHtml()`
+    assertion method for the `TestTrait`;
+* `Apache` class is now deprecated and will be removed in a later version.
+
 ### 1.1.12
 * added `is_iterable()` global function;
 * `assertIsArray()`, `assertIsInt()`, `assertIsObject()` and `assertIsString()`